refactor daemon run routes
